requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-openai
openai
qdrant-client
